Update patch set 1
Patch Set 1:

So you're just cherry-picking a single commit from a newer version of the NDK? Can we instead roll the entire NDK to a later version that includes the patch so we don't have to maintain local edits?

Also, just to clarify: this is code that would be staticly linked into libchromeview.so and therefore doesn't depend on specific versions of Android release, right?

Patch-set: 1
Label: Verified=0