Update patch set 1
Patch Set 1: Looks good to me, approved

(1 inline comment)

Ok, understood. This seems to match what's in the NDK so approving on that basis, but I'm hoping we can revert and roll to a newer NDK when it exists. It's mildly concerning that we're ahead of the game here and may discover new issues a result. +Grace as FYI

Patch-set: 1
Label: Code-Review=+2