Update patch set 1
Patch Set 1:

One more thing: the general policy for rolling NDK is that we only accept patches that are an official release, or in some cases we'll take a preview of a patch that has already landed in the aosp ndk, but which isn't part of an official release yet. Looking at the history, I see three commits related to crazy_linker that were added after r8e. Are all of those included in this release? Looks like r9b might include them but want to confirm.

Patch-set: 1