Update patch set 1
Patch Set 1:

No, it's the opposite, in fact.
The builtin definition for __builtin_isnan is wrong in clang. I suppose the NDK people thought it would be fixed in clang 3.5, but it isn't. What these definitions do is add attributes to these builtins for ARM.
So basically, __builtin_isnan is skipped for clang < 3.5, and I changed it to clang < 3.6.
It doesn't matter anyway because we don't use them in the first place, but it prevents the clang build from compiling.

Patch-set: 1
Label: Code-Review=0
Label: Commit-Queue=0