Update patch set 1
Patch Set 1:

In the newer NDK, most of the STL is defined with __attribute__((pcs("aapcs"))).
The issue here is that clang fails to resolve the template parameters when we overload the << operator in gtest.
This operator overload is not defined with the attribute and as a result, clang fails to see some templates are the same.
This patch restores the old behaviour, which is to define the STL functions without the attribute for clang.
This basically fixes this error:

../../third_party/WebKit/Source/web/tests/TouchActionTest.cpp:253:20: error: reference to overloaded function could not be resolved; did you mean to call it?
                << std::endl << "Document render tree:" << std::endl << externalRepresentation(root->document().frame()).utf8().data();
                   ^~~~~~~~~
/mnt/data/b/build/slave/android_clang_dbg/build/src/third_party/android_tools/ndk//sources/cxx-stl/stlport/stlport/stl/_ostream.h:357:1: note: possible target for call
endl(basic_ostream<_CharT, _Traits>& __os) {
^

(And I swear, I got it right, this time)

Patch-set: 1
Reviewer: Gerrit User 1116524 <1116524@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0