Remove .pyc files from ndk_experimental and fixing .gitignore
The ndk is checking in python and some libraries without excluding
pyc files. This causes the project to end up in an unclean state as
soon as the pyc files get recompiled and cause problems on sync.
pyc files should not be checked in. This change removes them and
excludes .pyc files in .gitignore.

BUG=354405