Remove .pyc files from ndk_experimental and fix .gitignore
Both ndk and ndk_experimental are checking in python+libs without
excluding pyc files. This causes the project to end up in an unclean
state as soon as the pyc files get recompiled causing problems on sync.
pyc files should not be checked in. This change removes them and
excludes .pyc files in .gitignore.

BUG=354405