Build libstlport with -fno-exceptions
The EH code of gabi++ doesn't unwind the stack frame built with -fexceptions.
This prevents DDMS native heap profiler to get stack traces of 'new' operator.
(The default 'new' operator is implemented in libstlport).

The fix is not to build libstlport with exception support, Chrome doesn't
support C++ exceptions anyway.

After changing the build script, use the following command to re-build
libstlport_shared.so and libstlport_static.a:

1) Download and extract NDK r9b rc2 into $HOME/android-ndk-r9b-rc2
2) export NDK=$HOME/android-ndk-r9b-rc2
3) ./build/tools/build-cxx-stl.sh --stl=stlport -j32 --abis=armeabi-v7a --ndk-dir=$NDK --gcc-version=4.8 --verbose