Update patch set 1
Patch Set 1: Looks good to me, approved

LGTM.
Just adding some more context about the libc.so replacement.

The libc.so file present in the NDK (wherever it comes from) is not shipped in the Chrome apk.
It merely serves as a compile/link-time stub to make the dynamic linker happy and generate the correct list of imported symbols into our chromeview.so.
Those symbols will be resolved at runtime (while loading the chromeview.so library) against the actual libc.so present on the device.

In essence what I'm saying is that replacing the libc.so LGTM. It doesn't really matter (it could contain even garbage instructions inside) as long as it exposes the right symbols.

Patch-set: 1
Label: Code-Review=+2
Label: Commit-Queue=0