Update patch set 1
Patch Set 1:

Considering this change was accepted upstream, I have no problem with it.

However, I do not understand why this fixes the Clang -Oz situation or if it is the proper fix for it. From my understanding, this is going to force the compiler to not generate standalone code for these functions. Question is, why does it work with -Os and not with -Oz?

Either way, please reference this change in README.chromium. Thanks!

Patch-set: 1
Label: Verified=0