Update patch set 1
Patch Set 1:

Good point Aurimas! The design support lib is distributed as a single jar(204kb), and TextInputLayout.class is only 12kb before compressing.  I remember last time I included recyclerview, the jar was 254kb.

I wonder if proguard is able to get rid of the resources that we don't actually use? If so the actual increase would be close to nothing.

Patch-set: 1