Update patch set 1
Patch Set 1:

Between that, cardboard and the next GMS we'd be hosed. Is there a logical subset of the library that we're using? If so you can pre-proguard that? Note that I don't want to start doing this at a per-class level. We should be looking for logical components that we're using vs the whole thing.
Sample CL from today to give you an idea:
https://chrome-internal-review.googlesource.com/#/c/219790/

Patch-set: 1