Update patch set 1
Patch Set 1:

Before landing this, could you check how many methods it adds to the APK in Debug mode?

  head -c 92 out/Debug/gen/clank/java/chrome_apk/classes.dex | tail -c 4 | hexdump -e '1/4 "%d\n"'

If we go over the 65K method limit (which we're dangerously close to), the build fails and the only solution is to start deleting third-party libraries or make proguard more aggressive. Not a good situation to be in! So, if this adds too many methods, we may want to consider other options. Hopefully that's not the case.

Patch-set: 1