Update patch set 1
Patch Set 1:

We don't *need* the cards library for now, but we might for the real feature. It's pretty simple to just swap the 9 patch in for the cardview, so I'll do that. bg_find_toolbar_popup is close -- it doesn't look exactly right, but it's close enough for this prototype. I'll abandon this change then and update the other CL.

Patch-set: 1
Reviewer: Gerrit User 1132179 <1132179@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0