Update patch set 1
Patch Set 1:

Khasim is trying to get the arm64 build to work with clang, which we don't currently have a bot for and it looks like the NDK headers are broken. Does this look okay? If so we should land this and then chase up Android about fixing the actual NDK.

Patch-set: 1