Update patch set 1
Patch Set 1:

From my point of view it seems reasonable, but the usual approach is to land a change in the Android AOSP repo first and then cherry-pick it here (with a note in README.chromium) so that it doesn't end up getting eradicated on the next NDK update. Can we do that or is this urgent?

Patch-set: 1