Update patch set 1
Patch Set 1:

I'm not talking about fixing an older NDK release, I'm talking about fixing the *next* NDK release so that it makes this change in all versions of headers. It doesn't appear that this change relies on an actual change to the platform libraries (since it's using an intrinsic, not a library function), so it seems like there's no particular reason why a binary compiled for platform-21 shouldn't also get the same fix.

Patch-set: 1