Update patch set 1
Patch Set 1:

Ah, I see it has; Khasim linked to the platform-24 version, which I took to mean that the platform-21 version had *not* been fixed, but it actually has.

So yes, we just need to roll to NDK r11, and if we aren't in a position to do that right now (not up to me) then we can safely land this change locally to fix it until we do, I think.

Patch-set: 1
Reviewer: Gerrit User 1109505 <1109505@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0