NDK: cherry-pick fix to make __func__ C99 compliant.
In the context of crrev.com/2161193003 it emerged that the Android
NDK mistakenly redefines __func__ = __PRETTY_FUNCTION__. This is
unneeded vestigial code. These days __func__ is a compiler builtin.
In practice this bug was blocking the C++11 feature rollout and
specifically the attempt of replacing __FUNCTION__ -> __func__.
This is a partial cherry-pick of:
https://android-review.googlesource.com/#/c/250323/2

BUG=

Change-Id: I68f6e8b3752cc97a68a85418d4a0cac5016d5f13
Reviewed-on: https://chromium-review.googlesource.com/363320
Reviewed-by: John Budorick <jbudorick@chromium.org>
Tested-by: John Budorick <jbudorick@chromium.org>