Update patch set 1
Patch Set 1:

> Is this even needed? We should not proguard gmscore upstream.
 > (build file: https://cs.chromium.org/chromium/src/build/secondary/third_party/android_tools/BUILD.gn?)
 >

Yes, this change is needed, but so is the secondary
BUILD.gn; however the latter can be done together
with the code and in a Chromium CL (FTR I have a
playground in https://crrev.com/2512123002 which
I'm landing in pieces).
 
 > There are some files here that I should probably delete.
 > 
 > What's the policy with the old unused gyp files? When can we delete
 > them?

Gyp should be completely unused in Chromium now
(see, e.g. the conversation in [1]).

[1] https://groups.google.com/a/chromium.org/d/msg/chromium-dev/RkeXrVmfXTQ/M7cJJ55pCQAJ

Patch-set: 1